test(select): migrate legacy wrapping
