+++ core/src/components/select/test/legacy/wrapping/select.e2e.ts-snapshots/select-nowrap-ios-ltr-Mobile-Chrome-linux.png
