+++ core/src/components/select/test/legacy/wrapping/select.e2e.ts-snapshots/select-nowrap-md-ltr-Mobile-Safari-linux.png
