+++ core/src/components/select/test/legacy/wrapping/select.e2e.ts-snapshots/select-wrap-ios-ltr-Mobile-Firefox-linux.png
