+++ core/src/components/select/test/legacy/wrapping/select.e2e.ts-snapshots/select-wrap-with-label-ios-ltr-Mobile-Safari-linux.png
